Compiling successful after refactor, tests broken
