+++ src/commands/mod.rs
@@ -1,8 +1,14 @@
 mod help;
 mod init;
+mod complete;
+mod execute;
 mod script;
+mod types;
 #[cfg(test)]
 mod tests;
 pub use self::help::help;
 pub use self::init::init;
+pub use self::complete::complete;
+pub use self::execute::execute;
 pub use self::script::Script;
+pub use self::types::{TargetType, CommandType};
