+++ src/commands/types.rs
@@ -1,0 +1,10 @@
+pub enum CommandType {
+    Execute,
+    Completion,
+}
+
+pub enum TargetType {
+    File,
+    Directory,
+}
+
