+++ src/tests.rs
@@ -1,6 +1,6 @@
 use super::execute;
 
-const EXAMPLE_DIR: &'static str = "./example";
+const EXAMPLE_DIR: &'static str = "--directory=./example";
 
 fn _vec_str(args: Vec<&str>) -> Vec<String> {
     args.iter().map(|s| s.to_string()).collect()
@@ -11,7 +11,7 @@ fn _vec_str(args: Vec<&str>) -> Vec<String> {
 #[test]
 fn test_simple_script() {
     assert_eq!(
-        execute(_vec_str(vec!["tome", EXAMPLE_DIR, "file_example"])),
+        execute(_vec_str(vec!["tome", "exec", EXAMPLE_DIR, "--", "file_example"])),
         Ok(format!("'{}/file_example'", EXAMPLE_DIR))
     );
 }
@@ -21,8 +21,9 @@ fn test_simple_script_completion() {
     assert_eq!(
         execute(_vec_str(vec![
             "tome",
+            "complete",
             EXAMPLE_DIR,
-            "--complete",
+            "--",
             "file_example",
         ])),
         Ok(String::from("file autocomplete example"))
@@ -33,7 +34,7 @@ fn test_simple_script_completion() {
 #[test]
 fn test_source() {
     assert_eq!(
-        execute(_vec_str(vec!["tome", EXAMPLE_DIR, "source_example",])),
+        execute(_vec_str(vec!["tome", "exec", EXAMPLE_DIR, "--", "source_example",])),
         Ok(format!("'.' '{}/source_example' ''", EXAMPLE_DIR))
     );
 }
@@ -44,8 +45,9 @@ fn test_source_completion() {
     assert_eq!(
         execute(_vec_str(vec![
             "tome",
+            "complete",
             EXAMPLE_DIR,
-            "--complete",
+            "--",
             "source_example",
         ])),
         Ok(String::from("foo baz\n"))
@@ -59,8 +61,9 @@ fn test_directory_completion() {
     assert_eq!(
         execute(_vec_str(vec![
             "tome",
+            "complete",
             EXAMPLE_DIR,
-            "--complete",
+            "--",
             "dir_example",
         ])),
         Ok("bar foo".to_string())
@@ -71,7 +74,7 @@ fn test_directory_completion() {
 #[test]
 fn test_root_directory_completion() {
     assert_eq!(
-        execute(_vec_str(vec!["tome", EXAMPLE_DIR, "--complete"])),
+        execute(_vec_str(vec!["tome", "complete", EXAMPLE_DIR])),
         Ok("dir_example file_example practical_examples source_example use-arg".to_string())
     );
 }
@@ -81,7 +84,7 @@ fn test_root_directory_completion() {
 #[test]
 fn test_script_in_directory() {
     assert_eq!(
-        execute(_vec_str(vec!["tome", EXAMPLE_DIR, "dir_example", "foo"])),
+        execute(_vec_str(vec!["tome", "complete", EXAMPLE_DIR, "--", "dir_example", "foo"])),
         Ok(format!("'{}/dir_example/foo'", EXAMPLE_DIR))
     );
 }
@@ -93,7 +96,9 @@ fn test_script_in_directory_not_found() {
     assert_eq!(
         execute(_vec_str(vec![
             "tome",
+            "exec",
             EXAMPLE_DIR,
+            "--",
             "dir_example",
             "foo-nonexistent",
             "baz"
@@ -109,7 +114,7 @@ fn test_script_in_directory_not_found() {
 #[test]
 fn test_script_directory_argument() {
     assert_eq!(
-        execute(_vec_str(vec!["tome", EXAMPLE_DIR, "dir_example",])),
+        execute(_vec_str(vec!["tome", "complete", EXAMPLE_DIR, "--", "dir_example",])),
         Err(format!(
             "{}/dir_example is a directory. tab-complete to choose subcommands",
             EXAMPLE_DIR
@@ -123,7 +128,7 @@ fn test_script_directory_argument() {
 #[test]
 fn test_use_arg() {
     assert_eq!(
-        execute(_vec_str(vec!["tome", EXAMPLE_DIR, "use-arg"])),
+        execute(_vec_str(vec!["tome", "exec", EXAMPLE_DIR, "--", "use-arg"])),
         Ok(format!("'.' '{}/use-arg' ''", EXAMPLE_DIR))
     );
 }
@@ -133,7 +138,7 @@ fn test_use_arg() {
 #[test]
 fn test_dangerous_characters_quoted() {
     assert_eq!(
-        execute(_vec_str(vec!["tome", EXAMPLE_DIR, "use-arg"])),
+        execute(_vec_str(vec!["tome", "exec", EXAMPLE_DIR, "--", "use-arg"])),
         Ok(format!("'.' '{}/use-arg' ''", EXAMPLE_DIR))
     );
 }
@@ -141,7 +146,7 @@ fn test_dangerous_characters_quoted() {
 /// help should be returned in no arguments are passed
 #[test]
 fn test_help_page() {
-    let result = execute(_vec_str(vec!["tome", EXAMPLE_DIR])).unwrap();
+    let result = execute(_vec_str(vec!["tome"])).unwrap();
     println!("{}", result);
     assert_eq!(result.matches("'\\''").count(), 1);
     assert_eq!(result.matches("'").count(), 5);
